Add a common function used to print formatted floats in the OSD Updated the OSD element and stats rendering code to use this fuction to print floating point values of varying precisions.
Will make adding OSD element variants that provide different decimal precisions (like voltage, amperage, etc.) easy to implement since all that needs to change is the `decimalPlaces` passed to the function. As an example, see the `osdFormatAltitudeString()` function.

If at some later date a more capable `printf` package is inplemented, then the guts of the `osdPrintFloat()` function could just be simplified while leaving the abstraction in place.

+++ src/main/osd/osd_elements.h
@@ -49,7 +49,7 @@ void osdFormatDistanceString(char *result, int distance, char leadingSymbol);
 bool osdFormatRtcDateTime(char *buffer);
 void osdFormatTime(char * buff, osd_timer_precision_e precision, timeUs_t time);
 void osdFormatTimer(char *buff, bool showSymbol, bool usePrecision, int timerIndex);
-int32_t osdGetMetersToSelectedUnit(int32_t meters);
+float osdGetMetersToSelectedUnit(int32_t meters);
 char osdGetMetersToSelectedUnitSymbol(void);
 int32_t osdGetSpeedToSelectedUnit(int32_t value);
 char osdGetSpeedToSelectedUnitSymbol(void);
